requirements.txt
asgiref==3.10.0
asttokens==3.0.0
attrs==25.4.0
black==25.9.0
certifi==2025.10.5
cfgv==3.4.0
charset-normalizer==3.4.4
click==8.3.0
coverage==7.11.0
decorator==5.2.1
distlib==0.4.0
Django==5.2.7
django-cors-headers==4.9.0
django-environ==0.12.0
django-filter==24.3
django-prometheus==2.4.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.1
dotenv==0.9.9
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.10.1
executing==2.2.1
filelock==3.20.0
flake8==7.3.0
gunicorn==23.0.0
identify==2.6.15
idna==3.11
inflection==0.5.1
iniconfig==2.1.0
ipython==9.6.0
ipython_pygments_lexers==1.1.1
isort==7.0.0
jedi==0.19.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
markdown-it-py==4.0.0
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
model-bakery==1.20.5
mypy_extensions==1.1.0
nodeenv==1.9.1
packaging==25.0
parso==0.8.5
pathspec==0.12.1
pexpect==4.9.0
pillow==12.0.0
platformdirs==4.5.0
pluggy==1.6.0
pre_commit==4.3.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
psycopg==3.2.10
psycopg-binary==3.2.10
psycopg2-binary==2.9.11
ptyprocess==0.7.0
pure_eval==0.2.3
python-dotenv>=1.0
pycodestyle==2.14.0
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.2
pytest-cov==7.0.0
pytest-django==4.11.1
python-dotenv==1.1.1
pytokens==0.2.0
PyYAML==6.0.3
RapidFuzz==3.14.1
referencing==0.37.0
requests==2.32.5
rich==14.2.0
rpds-py==0.27.1
sqlparse==0.5.3
stack-data==0.6.3
traitlets==5.14.3
uritemplate==4.2.0
urllib3==2.5.0
virtualenv==20.35.3
wcwidth==0.2.14
whitenoise==6.11.0
